Move and rename Library Service to api
Collect all domain incoming & outgoing traits
in one module.
