+++ incoming/http-api/src/lib.rs
@@ -1,8 +1,8 @@
 use crate::router::{AppState, OpenApiRouter};
 use axum::extract::DefaultBodyLimit;
+use pantsu_domain::api::incoming::ImageManagementService;
 use pantsu_domain::common::error::Error;
 use pantsu_domain::common::result::Result;
-use pantsu_domain::api::ImageManagementService;
 use pantsu_domain::reverse_image_search::ReverseImageSearchService;
 use pantsu_openapi::server;
 use std::sync::Arc;
