+++ outgoing/fs/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pantsu-fs-library"
+name = "pantsu-fs"
 version = "0.1.0"
 edition = "2024"
