+++ outgoing/fs/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod library;
+pub mod fs_image_repository;
