--- outgoing/fs_library/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod library;
-pub mod fs_service;
