+++ pantsu_server/Cargo.toml
@@ -16,7 +16,7 @@ path = "src/lib/lib.rs"
 [dependencies]
 pantsu-domain = { workspace = true }
 pantsu-http-api = { workspace = true }
-pantsu-fs-library = { workspace = true }
+pantsu-fs = { workspace = true }
 pantsu-iqdb = { workspace = true }
 pantsu-worker = { workspace = true }
